Task 16, first 98 fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,13 +9,13 @@
 int main(void)
 {
 	int i = 2;
-	unsigned long int fib_1 = 0;
-	unsigned long int fib_2 = 1;
-	unsigned long int fib_n = fib_1 + fib_2;
+	double fib_1 = 0;
+	double fib_2 = 1;
+	double fib_n = fib_1 + fib_2;
 
 	while (i <= 98)
 	{
-		printf("%lu, ", fib_n);
+		printf("%.0lf, ", fib_n);
 		fib_1 = fib_2;
 		fib_2 = fib_n;
 		fib_n = fib_1 + fib_2;
@@ -23,7 +23,7 @@ int main(void)
 		i++;
 	}
 	fib_n = fib_1 + fib_2;
-	printf("%lu\n", fib_n);
+	printf("%.0lf\n", fib_n);
 
 	return (0);
 }
